+++ src/util/addStaticDirToProxy.js
@@ -4,7 +4,7 @@ const path = require('path');
 
 // these routes are reserved by hammerhead and rammerhead
 const forbiddenRoutes = [
-    '/rammerhead.min.js',
+    '/rammerhead.js',
     '/hammerhead.js',
     '/task.js',
     '/iframe-task.js',